+++ heavenLi_pyopengl/drawUtils/drawHalo.cpp
@@ -10,12 +10,12 @@ int drawHalo(
       float bsy,                 /* y-Scale 2.0=spans display before GL scaling */
       float rs,                  /* Halo thickness */
       char circleSegments,       /* Number of sides */
-      double *color,             /* Polygon Color */
+      float *color,             /* Polygon Color */
       std::vector<float> &verts, /* Input Vector of x,y coordinates */
-      std::vector<double> &colrs  /* Input Vector of r,g,b values */
+      std::vector<float> &colrs  /* Input Vector of r,g,b values */
       ){
    float tma; 
-   double R, G, B;
+   float R, G, B;
    char degSegment = 360 / circleSegments;
    R = color[0];
    G = color[1];
